refactor: create separate module for github api type definitions
