+++ src/components/Repository/types.ts
@@ -1,4 +1,4 @@
-import { Repository } from "shared/types";
+import { Repository } from "shared/githubAPI";
 
 export interface RepositoryProps {
   repository: Repository;
